split database part from main.py
