+++ main.py
@@ -1,8 +1,11 @@
 from fastapi import FastAPI, Depends, HTTPException, Request
 from fastapi.responses import RedirectResponse
 from pydantic import BaseModel
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 
 import random, sqlite3, os
+
+import url_db, user_db
 
 class URLAdminInfo (BaseModel):
     original_url: str
@@ -11,51 +14,15 @@ class URLAdminInfo (BaseModel):
 class URLInfo (URLAdminInfo):
     clicks: int
 
-db_path = "url_data.db"
-
-def build_db():
-    if os.path.exists (db_path):
-        return
-    conn = sqlite3.connect (db_path)
-    cursor = conn.cursor()
-    cursor.execute ('''CREATE TABLE IF NOT EXISTS url_mapping (code TEXT PRIMARY KEY, url TEXT, clicks INTEGER DEFAULT 0)''')
-    conn.commit()
-    conn.close()
-
-def check_admin_code_is_available (admin_url: str):
-    conn = sqlite3.connect (db_path)
-    cursor = conn.cursor()
-    cursor.execute ('''SELECT code FROM url_mapping WHERE code = ?''', (admin_url,))
-    result = cursor.fetchone()
-    conn.close()
-
-    return result is None
-
-def insert_url (admin_url: str, ori: str):
-    if not check_admin_code_is_available (admin_url):
-        raise ValueError ("The chosen short code is already taken. Please choose a different one.")
-
-    conn = sqlite3.connect (db_path)
-    cursor = conn.cursor()
-    cursor.execute ('''INSERT INTO url_mapping (code, url) VALUES (?, ?)''', (admin_url, ori))
-    conn.commit()
-    conn.close()
-
-def update_click_count (admin_url: str):
-    conn = sqlite3.connect (db_path)
-    cursor = conn.cursor()
-    cursor.execute ('''UPDATE url_mapping SET clicks = clicks + ? WHERE code = ?''', (1, admin_url))
-    conn.commit()
-    conn.close()
-
-build_db()
+user_db.build_user_db()
+url_db.build_url_db()
 app = FastAPI()
 
 def raise_bad_request (message):
     raise HTTPException (status_code = 400, detail = message)
 
 def query (url: str):
-    conn = sqlite3.connect (db_path)
+    conn = sqlite3.connect (url_db.url_db_path)
     cursor = conn.cursor()
     cursor.execute ('''SELECT url FROM url_mapping WHERE code = ?''', (url,))
     result = cursor.fetchone()
@@ -99,7 +66,7 @@ async def get_url (url: str):
 
 @app.get ("/list/")
 async def list_url ():
-    conn = sqlite3.connect (db_path)
+    conn = sqlite3.connect (url_db.url_db_path)
     cursor = conn.cursor()
     cursor.execute ('''SELECT code, url, clicks FROM url_mapping''')
     all_mappings = cursor.fetchall()
@@ -112,7 +79,7 @@ async def delete_url (url: str):
     if check_admin_code_is_available (url):
         return "This admin url is not exist.\n"
 
-    conn = sqlite3.connect (db_path)
+    conn = sqlite3.connect (url_db.url_db_path)
     cursor = conn.cursor()
     cursor.execute ('''DELETE FROM url_mapping WHERE code = ?''', (url,))
     conn.commit()
